Day06/src/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
flake8==6.1.0
greenlet==3.0.1
grpcio==1.59.2
grpcio-tools==1.59.2
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.6.1
mypy-extensions==1.0.0
protobuf==4.25.0
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
tomli==2.0.1
types-protobuf==4.24.0.4
types-psycopg2==2.9.21.15
typing_extensions==4.8.0
